# grpc

+++ grpc-server/src/main/java/com/lzumetal/springboot/grpc/server/dao/BookDao.java
@@ -13,5 +13,8 @@ public interface BookDao {
 
 
 
-    Book selectById(int id);
+    Book selectById(Long id);
+
+
+
 }
